Working on user display bug
Code is now working when the user signs in with a Google account. Fixed
up session dataabse code so that way the user will stay signed in if you
were to reload the website. useful during testing. This bug friggn sux
and ive spent to many hours reading github issue threads. have a good
day <3